image_generator/pyproject.toml
[project]
name = "image-generator"
version = "0.1.0"
description = "Geracao de imagens"
authors = [
    {name = "Junior Torres"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai (>=1.75.0,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "schedule (>=1.2.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
